fix: 🔧 move test utility to test only
